[build] as suggested via email, beautify and fix a subshell issue in make_pyanaconda_tarball.sh

+++ make_pyanaconda_tarball.sh
@@ -5,17 +5,17 @@ mkdir tmp
 sh autogen.sh
 ./configure --prefix=/usr --disable-selinux --enable-selinux=no
 make
-make DESTDIR=$PWD/tmp install
-find $PWD/tmp -name "*.py[co]" | xargs rm
-find $PWD/tmp -name "*.*a" | xargs rm
-find $PWD/tmp -name "*.so" | xargs rm
+make DESTDIR="${PWD}"/tmp install
+find "${PWD}"/tmp -name "*.py[co]" | xargs rm
+find "${PWD}"/tmp -name "*.*a" | xargs rm
+find "${PWD}"/tmp -name "*.so" | xargs rm
 pyanaconda_dir=$(find tmp -name "pyanaconda" -type d)
-[[ -d "${pyanaconda_dir}" ]] || (echo "ouch" && exit 1)
-cd $(dirname ${pyanaconda_dir})
+[[ -d "${pyanaconda_dir}" ]] || { echo "ouch" && exit 1 }
+cd $(dirname "${pyanaconda_dir}")
 tar cjvf pyanaconda-${SABAYON_VER}.tar.bz2 "$(basename ${pyanaconda_dir})"
 md5sum pyanaconda-${SABAYON_VER}.tar.bz2 > pyanaconda-${SABAYON_VER}.tar.bz2.md5
-mv pyanaconda-${SABAYON_VER}.tar.bz2{,.md5} $OLDPWD/
-cd $OLDPWD
+mv pyanaconda-${SABAYON_VER}.tar.bz2{,.md5} "${OLDPWD}"/
+cd "${OLDPWD}"
 make distclean
 rm -rf tmp
 echo
